requirements.txt
asgiref==3.3.4
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.2
django-pandas==0.6.4
flatlib==0.2.3
googlemaps==4.4.5
gunicorn==20.1.0
html5lib==1.1
idna==2.10
numpy==1.20.3
odfpy==1.4.1
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
psycopg2-binary==2.9.1
pycparser==2.20
Pyphen==0.10.0
pyswisseph==2.8.0.post1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.4
webencodings==0.5.1
